# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API WitelonBank",
        "description": "Dokumentacja API dla systemu bankowego WitelonBank. Pamiętaj o autoryzacji (przycisk 'Authorize' na górze) dla chronionych endpointów.",
        "contact": {
            "email": "support@witelonbank.com"
        },
        "license": {
            "name": "MIT License",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Główny serwer API"
        }
    ],
    "paths": {
...
        "/api/kryptowaluty/ceny": {
            "get": {
                "tags": [
                    "Inwestycje"
                ],
                "summary": "Pobiera listę dostępnych kryptowalut i ich aktualne ceny z API",
                "operationId": "499eabc1fc36915ba7d0ce84ccd98766",
                "responses": {
                    "200": {
                        "description": "Lista kryptowalut z cenami",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "BTC": 250000.5,
                                    "ETH": 12000.75
                                }
                            }
                        }
                    },
                    "503": {
                        "description": "Nie udało się pobrać cen z zewnętrznego API",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Nie udało się pobrać aktualnych cen kryptowalut."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/inwestycje/kup": {
            "post": {
                "tags": [
                    "Inwestycje"
                ],
                "summary": "Kupuje kryptowalutę za środki z konta PLN",
                "operationId": "ed1499ed6038a4369e1a8bc20791ff74",
                "requestBody": {
                    "description": "Dane do zakupu kryptowaluty",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_konta_pln",
                                    "symbol_krypto",
                                    "kwota_pln"
                                ],
                                "properties": {
                                    "id_konta_pln": {
                                        "description": "ID konta PLN, z którego pobierane są środki",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "symbol_krypto": {
                                        "description": "Symbol kryptowaluty do kupienia",
                                        "type": "string",
                                        "enum": [
                                            "BTC",
                                            "ETH"
                                        ],
                                        "example": "BTC"
                                    },
                                    "kwota_pln": {
                                        "description": "Kwota w PLN do zainwestowania",
                                        "type": "number",
                                        "format": "float",
                                        "example": 100
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Kryptowaluta zakupiona pomyślnie",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Zakupiono X BTC za Y PLN."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Błąd walidacji lub niewystarczające środki",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorValidation"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Konto PLN lub portfel nie znalezione"
                    },
                    "503": {
                        "description": "Nie udało się pobrać aktualnej ceny kryptowaluty z API"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/portfel": {
            "get": {
                "tags": [
                    "Inwestycje"
                ],
                "summary": "Pobiera aktualny portfel kryptowalut zalogowanego użytkownika",
                "operationId": "187222d27d38613f116b18b1707b0db2",
                "responses": {
                    "200": {
                        "description": "Portfel użytkownika"
                    },
                    "404": {
                        "description": "Portfel nie znaleziony (użytkownik jeszcze nie inwestował)",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Nie znaleziono portfela dla tego użytkownika."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ErrorValidation": {
                "title": "Błąd Walidacji",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Podane dane są nieprawidłowe."
                    },
                    "errors": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "array",
                            "items": {
                                "type": "string",
                                "example": "Pole email jest wymagane."
                            }
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Wprowadź token JWT z prefiksem Bearer (np. 'Bearer eyJhbGciOiJIUzI1NiI...). Token uzyskasz po zalogowaniu.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}